storage/cloud: split out implementations into own files
This is change is only moving code, and only what files that code is in -- no behavior or even symbols are changed.
This is just splitting up the big 'external_storage.go' that previously contained both the common interface and
all the implementaitons of it, to only contain the interface and common helpers, with each implementaiton now in its
own file. Tests are similarly split.

Release note: none.
