+++ pkg/storage/cloud/external_storage_test.go
@@ -15,28 +15,18 @@ import (
 	"context"
 	"crypto/rand"
 	"encoding/base64"
-	"encoding/pem"
 	"fmt"
-	"io"
 	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
 	"net/url"
 	"os"
 	"path/filepath"
 	"sort"
-	"strconv"
 	"strings"
 	"testing"
-	"time"
 
-	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/cockroachdb/cockroach/pkg/blobs"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
-	"github.com/cockroachdb/cockroach/pkg/testutils"
-	"github.com/cockroachdb/cockroach/pkg/util/ctxgroup"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
-	"github.com/cockroachdb/cockroach/pkg/util/retry"
 	"github.com/cockroachdb/cockroach/pkg/workload"
 	"github.com/cockroachdb/cockroach/pkg/workload/bank"
 	"github.com/spf13/pflag"
@@ -319,295 +309,6 @@ func testListFiles(t *testing.T, storeURI string) {
 	}
 }
 
-func TestPutLocal(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	p, cleanupFn := testutils.TempDir(t)
-	defer cleanupFn()
-
-	testSettings.ExternalIODir = p
-	dest := MakeLocalStorageURI(p)
-
-	testExportStore(t, dest, false)
-	testListFiles(t, fmt.Sprintf("nodelocal:///%s", "listing-test"))
-}
-
-func TestLocalIOLimits(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	ctx := context.TODO()
-	const allowed = "/allowed"
-	testSettings.ExternalIODir = allowed
-
-	clientFactory := blobs.TestBlobServiceClient(testSettings.ExternalIODir)
-	for dest, expected := range map[string]string{allowed: "", "/../../blah": "not allowed"} {
-		u := fmt.Sprintf("nodelocal://%s", dest)
-		e, err := ExternalStorageFromURI(ctx, u, testSettings, clientFactory)
-		if err != nil {
-			t.Fatal(err)
-		}
-		_, err = e.ListFiles(ctx)
-		if !testutils.IsError(err, expected) {
-			t.Fatal(err)
-		}
-	}
-
-	for host, expectErr := range map[string]bool{"": false, "1": false, "0": false, "blah": true} {
-		u := fmt.Sprintf("nodelocal://%s/path/to/file", host)
-
-		var expected string
-		if expectErr {
-			expected = "host component of nodelocal URI must be a node ID"
-		}
-		if _, err := ExternalStorageConfFromURI(u); !testutils.IsError(err, expected) {
-			t.Fatalf("%q: expected error %q, got %v", u, expected, err)
-		}
-	}
-}
-
-func TestPutHttp(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	tmp, dirCleanup := testutils.TempDir(t)
-	defer dirCleanup()
-
-	const badHeadResponse = "bad-head-response"
-
-	makeServer := func() (*url.URL, func() int, func()) {
-		var files int
-		srv := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			localfile := filepath.Join(tmp, filepath.Base(r.URL.Path))
-			switch r.Method {
-			case "PUT":
-				f, err := os.Create(localfile)
-				if err != nil {
-					http.Error(w, err.Error(), 500)
-					return
-				}
-				defer f.Close()
-				if _, err := io.Copy(f, r.Body); err != nil {
-					http.Error(w, err.Error(), 500)
-					return
-				}
-				files++
-				w.WriteHeader(201)
-			case "GET", "HEAD":
-				if filepath.Base(localfile) == badHeadResponse {
-					http.Error(w, "HEAD not implemented", 500)
-					return
-				}
-				http.ServeFile(w, r, localfile)
-			case "DELETE":
-				if err := os.Remove(localfile); err != nil {
-					http.Error(w, err.Error(), 500)
-					return
-				}
-				w.WriteHeader(204)
-			default:
-				http.Error(w, "unsupported method "+r.Method, 400)
-			}
-		}))
-
-		u := testSettings.MakeUpdater()
-		if err := u.Set(
-			cloudstorageHTTPCASetting,
-			string(pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: srv.Certificate().Raw})),
-			"s",
-		); err != nil {
-			t.Fatal(err)
-		}
-
-		cleanup := func() {
-			srv.Close()
-			if err := u.Set(cloudstorageHTTPCASetting, "", "s"); err != nil {
-				t.Fatal(err)
-			}
-		}
-
-		t.Logf("Mock HTTP Storage %q", srv.URL)
-		uri, err := url.Parse(srv.URL)
-		if err != nil {
-			srv.Close()
-			t.Fatal(err)
-		}
-		uri.Path = filepath.Join(uri.Path, "testing")
-		return uri, func() int { return files }, cleanup
-	}
-
-	t.Run("singleHost", func(t *testing.T) {
-		srv, files, cleanup := makeServer()
-		defer cleanup()
-		testExportStore(t, srv.String(), false)
-		if expected, actual := 13, files(); expected != actual {
-			t.Fatalf("expected %d files to be written to single http store, got %d", expected, actual)
-		}
-	})
-
-	t.Run("multiHost", func(t *testing.T) {
-		srv1, files1, cleanup1 := makeServer()
-		defer cleanup1()
-		srv2, files2, cleanup2 := makeServer()
-		defer cleanup2()
-		srv3, files3, cleanup3 := makeServer()
-		defer cleanup3()
-
-		combined := *srv1
-		combined.Host = strings.Join([]string{srv1.Host, srv2.Host, srv3.Host}, ",")
-
-		testExportStore(t, combined.String(), true)
-		if expected, actual := 3, files1(); expected != actual {
-			t.Fatalf("expected %d files written to http host 1, got %d", expected, actual)
-		}
-		if expected, actual := 4, files2(); expected != actual {
-			t.Fatalf("expected %d files written to http host 2, got %d", expected, actual)
-		}
-		if expected, actual := 4, files3(); expected != actual {
-			t.Fatalf("expected %d files written to http host 3, got %d", expected, actual)
-		}
-	})
-
-	// Ensure that servers that error on HEAD are handled gracefully.
-	t.Run("bad-head-response", func(t *testing.T) {
-		ctx := context.TODO()
-
-		srv, _, cleanup := makeServer()
-		defer cleanup()
-
-		conf, err := ExternalStorageConfFromURI(srv.String())
-		if err != nil {
-			t.Fatal(err)
-		}
-		s, err := MakeExternalStorage(ctx, conf, testSettings, blobs.TestEmptyBlobClientFactory)
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer s.Close()
-
-		const file = "file"
-		var content = []byte("contents")
-		if err := s.WriteFile(ctx, file, bytes.NewReader(content)); err != nil {
-			t.Fatal(err)
-		}
-		if err := s.WriteFile(ctx, badHeadResponse, bytes.NewReader(content)); err != nil {
-			t.Fatal(err)
-		}
-		if sz, err := s.Size(ctx, file); err != nil {
-			t.Fatal(err)
-		} else if sz != int64(len(content)) {
-			t.Fatalf("expected %d, got %d", len(content), sz)
-		}
-		if sz, err := s.Size(ctx, badHeadResponse); !testutils.IsError(err, "500 Internal Server Error") {
-			t.Fatalf("unexpected error: %v", err)
-		} else if sz != 0 {
-			t.Fatalf("expected 0 size, got %d", sz)
-		}
-	})
-}
-
-func TestPutS3(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	// If environment credentials are not present, we want to
-	// skip all S3 tests, including auth-implicit, even though
-	// it is not used in auth-implicit.
-	creds, err := credentials.NewEnvCredentials().Get()
-	if err != nil {
-		t.Skip("No AWS credentials")
-	}
-	bucket := os.Getenv("AWS_S3_BUCKET")
-	if bucket == "" {
-		t.Skip("AWS_S3_BUCKET env var must be set")
-	}
-
-	ctx := context.TODO()
-	t.Run("auth-empty-no-cred", func(t *testing.T) {
-		_, err := ExternalStorageFromURI(
-			ctx, fmt.Sprintf("s3://%s/%s", bucket, "backup-test-default"),
-			testSettings, blobs.TestEmptyBlobClientFactory,
-		)
-		require.EqualError(t, err, fmt.Sprintf(
-			`%s is set to '%s', but %s is not set`,
-			AuthParam,
-			authParamSpecified,
-			S3AccessKeyParam,
-		))
-	})
-	t.Run("auth-implicit", func(t *testing.T) {
-		// You can create an IAM that can access S3
-		// in the AWS console, then set it up locally.
-		// https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-role.html
-		// We only run this test if default role exists.
-		credentialsProvider := credentials.SharedCredentialsProvider{}
-		_, err := credentialsProvider.Retrieve()
-		if err != nil {
-			t.Skip(err)
-		}
-
-		testExportStore(
-			t,
-			fmt.Sprintf(
-				"s3://%s/%s?%s=%s",
-				bucket, "backup-test-default",
-				AuthParam, authParamImplicit,
-			),
-			false,
-		)
-	})
-
-	t.Run("auth-specified", func(t *testing.T) {
-		testExportStore(t,
-			fmt.Sprintf(
-				"s3://%s/%s?%s=%s&%s=%s",
-				bucket, "backup-test",
-				S3AccessKeyParam, url.QueryEscape(creds.AccessKeyID),
-				S3SecretParam, url.QueryEscape(creds.SecretAccessKey),
-			),
-			false,
-		)
-		testListFiles(t,
-			fmt.Sprintf(
-				"s3://%s/%s?%s=%s&%s=%s",
-				bucket, "listing-test",
-				S3AccessKeyParam, url.QueryEscape(creds.AccessKeyID),
-				S3SecretParam, url.QueryEscape(creds.SecretAccessKey),
-			),
-		)
-	})
-}
-
-func TestPutS3Endpoint(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	q := make(url.Values)
-	expect := map[string]string{
-		"AWS_S3_ENDPOINT":        S3EndpointParam,
-		"AWS_S3_ENDPOINT_KEY":    S3AccessKeyParam,
-		"AWS_S3_ENDPOINT_REGION": S3RegionParam,
-		"AWS_S3_ENDPOINT_SECRET": S3SecretParam,
-	}
-	for env, param := range expect {
-		v := os.Getenv(env)
-		if v == "" {
-			t.Skipf("%s env var must be set", env)
-		}
-		q.Add(param, v)
-	}
-
-	bucket := os.Getenv("AWS_S3_ENDPOINT_BUCKET")
-	if bucket == "" {
-		t.Skip("AWS_S3_ENDPOINT_BUCKET env var must be set")
-	}
-
-	u := url.URL{
-		Scheme:   "s3",
-		Host:     bucket,
-		Path:     "backup-test",
-		RawQuery: q.Encode(),
-	}
-
-	testExportStore(t, u.String(), false)
-}
-
 func TestPutGoogleCloud(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 
@@ -664,37 +365,6 @@ func TestPutGoogleCloud(t *testing.T) {
 			false,
 		)
 	})
-}
-
-func TestPutAzure(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	accountName := os.Getenv("AZURE_ACCOUNT_NAME")
-	accountKey := os.Getenv("AZURE_ACCOUNT_KEY")
-	if accountName == "" || accountKey == "" {
-		t.Skip("AZURE_ACCOUNT_NAME and AZURE_ACCOUNT_KEY env vars must be set")
-	}
-	bucket := os.Getenv("AZURE_CONTAINER")
-	if bucket == "" {
-		t.Skip("AZURE_CONTAINER env var must be set")
-	}
-
-	testExportStore(t,
-		fmt.Sprintf("azure://%s/%s?%s=%s&%s=%s",
-			bucket, "backup-test",
-			AzureAccountNameParam, url.QueryEscape(accountName),
-			AzureAccountKeyParam, url.QueryEscape(accountKey),
-		),
-		false,
-	)
-	testListFiles(
-		t,
-		fmt.Sprintf("azure://%s/%s?%s=%s&%s=%s",
-			bucket, "listing-test",
-			AzureAccountNameParam, url.QueryEscape(accountName),
-			AzureAccountKeyParam, url.QueryEscape(accountKey),
-		),
-	)
 }
 
 func TestWorkloadStorage(t *testing.T) {
@@ -786,114 +456,4 @@ func TestWorkloadStorage(t *testing.T) {
 		ctx, `workload:///csv/bank/bank?version=nope`, settings, blobs.TestEmptyBlobClientFactory,
 	)
 	require.EqualError(t, err, `expected bank version "nope" but got "1.0.0"`)
-}
-
-func rangeStart(r string) (int, error) {
-	if len(r) == 0 {
-		return 0, nil
-	}
-	r = strings.TrimPrefix(r, "bytes=")
-
-	return strconv.Atoi(r[:strings.IndexByte(r, '-')])
-}
-
-func TestHttpGet(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	data := []byte("to serve, or not to serve.  c'est la question")
-
-	httpRetryOptions.InitialBackoff = 1 * time.Microsecond
-	httpRetryOptions.MaxBackoff = 10 * time.Millisecond
-	httpRetryOptions.MaxRetries = 100
-
-	for _, tc := range []int{1, 2, 5, 16, 32, len(data) - 1, len(data)} {
-		t.Run(fmt.Sprintf("read-%d", tc), func(t *testing.T) {
-			limit := tc
-			s := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				start, err := rangeStart(r.Header.Get("Range"))
-				if start < 0 || start >= len(data) {
-					t.Errorf("invalid start offset %d in range header %s",
-						start, r.Header.Get("Range"))
-				}
-				end := start + limit
-				if end > len(data) {
-					end = len(data)
-				}
-
-				w.Header().Add("Accept-Ranges", "bytes")
-				w.Header().Add("Content-Length", strconv.Itoa(len(data)-start))
-
-				if start > 0 {
-					w.Header().Add(
-						"Content-Range",
-						fmt.Sprintf("bytes %d-%d/%d", start, end, len(data)))
-				}
-
-				if err == nil {
-					_, err = w.Write(data[start:end])
-				}
-				if err != nil {
-					w.WriteHeader(http.StatusInternalServerError)
-				}
-			}))
-
-			// Start antagonist function that aggressively closes client connections.
-			ctx, cancelAntagonist := context.WithCancel(context.Background())
-			g := ctxgroup.WithContext(ctx)
-			g.GoCtx(func(ctx context.Context) error {
-				opts := retry.Options{
-					InitialBackoff: 500 * time.Microsecond,
-					MaxBackoff:     1 * time.Millisecond,
-				}
-				for attempt := retry.StartWithCtx(ctx, opts); attempt.Next(); {
-					s.CloseClientConnections()
-				}
-				return nil
-			})
-
-			store, err := makeHTTPStorage(s.URL, testSettings)
-			require.NoError(t, err)
-
-			var file io.ReadCloser
-
-			// Cleanup.
-			defer func() {
-				s.Close()
-				if store != nil {
-					require.NoError(t, store.Close())
-				}
-				if file != nil {
-					require.NoError(t, file.Close())
-				}
-				cancelAntagonist()
-				_ = g.Wait()
-			}()
-
-			// Read the file and verify results.
-			file, err = store.ReadFile(ctx, "/something")
-			require.NoError(t, err)
-
-			b, err := ioutil.ReadAll(file)
-			require.NoError(t, err)
-			require.EqualValues(t, data, b)
-		})
-	}
-}
-
-func TestHttpGetWithCancelledContext(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	s := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {}))
-	defer s.Close()
-
-	store, err := makeHTTPStorage(s.URL, testSettings)
-	require.NoError(t, err)
-	defer func() {
-		require.NoError(t, store.Close())
-	}()
-
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	_, err = store.ReadFile(ctx, "/something")
-	require.Error(t, context.Canceled, err)
 }
