+++ pkg/storage/cloud/gcs_storage.go
@@ -1,0 +1,210 @@
+// Copyright 2019 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package cloud
+
+import (
+	"context"
+	"encoding/base64"
+	"fmt"
+	"io"
+	"net/url"
+	"path/filepath"
+
+	gcs "cloud.google.com/go/storage"
+	"github.com/cockroachdb/cockroach/pkg/base"
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
+	"github.com/cockroachdb/cockroach/pkg/util/contextutil"
+	"github.com/cockroachdb/cockroach/pkg/util/retry"
+	"github.com/cockroachdb/errors"
+	"golang.org/x/oauth2/google"
+	"google.golang.org/api/iterator"
+	"google.golang.org/api/option"
+)
+
+type gcsStorage struct {
+	bucket   *gcs.BucketHandle
+	client   *gcs.Client
+	conf     *roachpb.ExternalStorage_GCS
+	prefix   string
+	settings *cluster.Settings
+}
+
+var _ ExternalStorage = &gcsStorage{}
+
+func (g *gcsStorage) Conf() roachpb.ExternalStorage {
+	return roachpb.ExternalStorage{
+		Provider:          roachpb.ExternalStorageProvider_GoogleCloud,
+		GoogleCloudConfig: g.conf,
+	}
+}
+
+func makeGCSStorage(
+	ctx context.Context, conf *roachpb.ExternalStorage_GCS, settings *cluster.Settings,
+) (ExternalStorage, error) {
+	if conf == nil {
+		return nil, errors.Errorf("google cloud storage upload requested but info missing")
+	}
+	const scope = gcs.ScopeReadWrite
+	opts := []option.ClientOption{option.WithScopes(scope)}
+
+	// "default": only use the key in the settings; error if not present.
+	// "specified": the JSON object for authentication is given by the CREDENTIALS param.
+	// "implicit": only use the environment data.
+	// "": if default key is in the settings use it; otherwise use environment data.
+	switch conf.Auth {
+	case "", authParamDefault:
+		var key string
+		if settings != nil {
+			key = gcsDefault.Get(&settings.SV)
+		}
+		// We expect a key to be present if default is specified.
+		if conf.Auth == authParamDefault && key == "" {
+			return nil, errors.Errorf("expected settings value for %s", cloudstorageGSDefaultKey)
+		}
+		if key != "" {
+			source, err := google.JWTConfigFromJSON([]byte(key), scope)
+			if err != nil {
+				return nil, errors.Wrap(err, "creating GCS oauth token source")
+			}
+			opts = append(opts, option.WithTokenSource(source.TokenSource(ctx)))
+		}
+	case authParamSpecified:
+		if conf.Credentials == "" {
+			return nil, errors.Errorf(
+				"%s is set to '%s', but %s is not set",
+				AuthParam,
+				authParamSpecified,
+				CredentialsParam,
+			)
+		}
+		decodedKey, err := base64.StdEncoding.DecodeString(conf.Credentials)
+		if err != nil {
+			return nil, errors.Wrap(err, fmt.Sprintf("decoding value of %s", CredentialsParam))
+		}
+		source, err := google.JWTConfigFromJSON(decodedKey, scope)
+		if err != nil {
+			return nil, errors.Wrap(err, "creating GCS oauth token source from specified credentials")
+		}
+		opts = append(opts, option.WithTokenSource(source.TokenSource(ctx)))
+	case authParamImplicit:
+		// Do nothing; use implicit params:
+		// https://godoc.org/golang.org/x/oauth2/google#FindDefaultCredentials
+	default:
+		return nil, errors.Errorf("unsupported value %s for %s", conf.Auth, AuthParam)
+	}
+	g, err := gcs.NewClient(ctx, opts...)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to create google cloud client")
+	}
+	bucket := g.Bucket(conf.Bucket)
+	if conf.BillingProject != `` {
+		bucket = bucket.UserProject(conf.BillingProject)
+	}
+	return &gcsStorage{
+		bucket:   bucket,
+		client:   g,
+		conf:     conf,
+		prefix:   conf.Prefix,
+		settings: settings,
+	}, nil
+}
+
+func (g *gcsStorage) WriteFile(ctx context.Context, basename string, content io.ReadSeeker) error {
+	const maxAttempts = 3
+	err := retry.WithMaxAttempts(ctx, base.DefaultRetryOptions(), maxAttempts, func() error {
+		if _, err := content.Seek(0, io.SeekStart); err != nil {
+			return err
+		}
+		// Set the timeout within the retry loop.
+		return contextutil.RunWithTimeout(ctx, "put gcs file", timeoutSetting.Get(&g.settings.SV),
+			func(ctx context.Context) error {
+				w := g.bucket.Object(filepath.Join(g.prefix, basename)).NewWriter(ctx)
+				if _, err := io.Copy(w, content); err != nil {
+					_ = w.Close()
+					return err
+				}
+				return w.Close()
+			})
+	})
+	return errors.Wrap(err, "write to google cloud")
+}
+
+func (g *gcsStorage) ReadFile(ctx context.Context, basename string) (io.ReadCloser, error) {
+	// https://github.com/cockroachdb/cockroach/issues/23859
+
+	var rc io.ReadCloser
+	err := delayedRetry(ctx, func() error {
+		var readErr error
+		rc, readErr = g.bucket.Object(filepath.Join(g.prefix, basename)).NewReader(ctx)
+		return readErr
+	})
+	return rc, err
+}
+
+func (g *gcsStorage) ListFiles(ctx context.Context) ([]string, error) {
+	var fileList []string
+	it := g.bucket.Objects(ctx, &gcs.Query{
+		Prefix: getBucketBeforeWildcard(g.prefix),
+	})
+	for {
+		attrs, err := it.Next()
+		if err == iterator.Done {
+			break
+		}
+		if err != nil {
+			return nil, errors.Wrap(err, "unable to list files in gcs bucket")
+		}
+
+		matches, errMatch := filepath.Match(g.prefix, attrs.Name)
+		if errMatch != nil {
+			continue
+		}
+		if matches {
+			gsURL := url.URL{
+				Scheme: "gs",
+				Host:   attrs.Bucket,
+				Path:   attrs.Name,
+			}
+			fileList = append(fileList, gsURL.String())
+		}
+	}
+
+	return fileList, nil
+}
+
+func (g *gcsStorage) Delete(ctx context.Context, basename string) error {
+	return contextutil.RunWithTimeout(ctx, "delete gcs file",
+		timeoutSetting.Get(&g.settings.SV),
+		func(ctx context.Context) error {
+			return g.bucket.Object(filepath.Join(g.prefix, basename)).Delete(ctx)
+		})
+}
+
+func (g *gcsStorage) Size(ctx context.Context, basename string) (int64, error) {
+	var r *gcs.Reader
+	if err := contextutil.RunWithTimeout(ctx, "size gcs file",
+		timeoutSetting.Get(&g.settings.SV),
+		func(ctx context.Context) error {
+			var err error
+			r, err = g.bucket.Object(filepath.Join(g.prefix, basename)).NewReader(ctx)
+			return err
+		}); err != nil {
+		return 0, err
+	}
+	sz := r.Attrs.Size
+	_ = r.Close()
+	return sz, nil
+}
+
+func (g *gcsStorage) Close() error {
+	return g.client.Close()
+}
