+++ pkg/storage/cloud/nodelocal_storage.go
@@ -1,0 +1,122 @@
+// Copyright 2019 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package cloud
+
+import (
+	"context"
+	"fmt"
+	"io"
+	"path"
+	"strings"
+
+	"github.com/cockroachdb/cockroach/pkg/blobs"
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
+	"github.com/cockroachdb/errors"
+)
+
+type localFileStorage struct {
+	cfg        roachpb.ExternalStorage_LocalFilePath // contains un-prefixed filepath -- DO NOT use for I/O ops.
+	base       string                                // relative filepath prefixed with externalIODir, for I/O ops on this node.
+	blobClient blobs.BlobClient                      // inter-node file sharing service
+}
+
+var _ ExternalStorage = &localFileStorage{}
+
+// MakeLocalStorageURI converts a local path (should always be relative) to a
+// valid nodelocal URI.
+func MakeLocalStorageURI(path string) string {
+	return fmt.Sprintf("nodelocal:///%s", path)
+}
+
+func makeNodeLocalURIWithNodeID(nodeID roachpb.NodeID, path string) string {
+	path = strings.TrimPrefix(path, "/")
+	if nodeID == 0 {
+		return fmt.Sprintf("nodelocal:///%s", path)
+	}
+	return fmt.Sprintf("nodelocal://%d/%s", nodeID, path)
+}
+
+func makeLocalStorage(
+	ctx context.Context,
+	cfg roachpb.ExternalStorage_LocalFilePath,
+	settings *cluster.Settings,
+	blobClientFactory blobs.BlobClientFactory,
+) (ExternalStorage, error) {
+	if cfg.Path == "" {
+		return nil, errors.Errorf("Local storage requested but path not provided")
+	}
+	client, err := blobClientFactory(ctx, cfg.NodeID)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to create blob client")
+	}
+
+	// In non-server execution we have no settings and no restriction on local IO.
+	if settings != nil {
+		if settings.ExternalIODir == "" {
+			return nil, errors.Errorf("local file access is disabled")
+		}
+	}
+	return &localFileStorage{base: cfg.Path, cfg: cfg, blobClient: client}, nil
+}
+
+func (l *localFileStorage) Conf() roachpb.ExternalStorage {
+	return roachpb.ExternalStorage{
+		Provider:  roachpb.ExternalStorageProvider_LocalFile,
+		LocalFile: l.cfg,
+	}
+}
+
+func joinRelativePath(filePath string, file string) string {
+	// Joining "." to make this a relative path.
+	// This ensures path.Clean does not simplify in unexpected ways.
+	return path.Join(".", filePath, file)
+}
+
+func (l *localFileStorage) WriteFile(
+	ctx context.Context, basename string, content io.ReadSeeker,
+) error {
+	return l.blobClient.WriteFile(ctx, joinRelativePath(l.base, basename), content)
+}
+
+func (l *localFileStorage) ReadFile(ctx context.Context, basename string) (io.ReadCloser, error) {
+	return l.blobClient.ReadFile(ctx, joinRelativePath(l.base, basename))
+}
+
+func (l *localFileStorage) ListFiles(ctx context.Context) ([]string, error) {
+	var fileList []string
+	matches, err := l.blobClient.List(ctx, l.base)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to match pattern provided")
+	}
+
+	for _, fileName := range matches {
+		fileList = append(fileList, makeNodeLocalURIWithNodeID(l.cfg.NodeID, fileName))
+	}
+
+	return fileList, nil
+}
+
+func (l *localFileStorage) Delete(ctx context.Context, basename string) error {
+	return l.blobClient.Delete(ctx, joinRelativePath(l.base, basename))
+}
+
+func (l *localFileStorage) Size(ctx context.Context, basename string) (int64, error) {
+	stat, err := l.blobClient.Stat(ctx, joinRelativePath(l.base, basename))
+	if err != nil {
+		return 0, err
+	}
+	return stat.Filesize, nil
+}
+
+func (*localFileStorage) Close() error {
+	return nil
+}
