+++ pkg/storage/cloud/s3_storage.go
@@ -1,0 +1,224 @@
+// Copyright 2019 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package cloud
+
+import (
+	"context"
+	"io"
+	"net/url"
+	"path/filepath"
+	"strings"
+
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/s3"
+	"github.com/aws/aws-sdk-go/service/s3/s3manager"
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
+	"github.com/cockroachdb/cockroach/pkg/util/contextutil"
+	"github.com/cockroachdb/errors"
+)
+
+type s3Storage struct {
+	bucket   *string
+	conf     *roachpb.ExternalStorage_S3
+	prefix   string
+	s3       *s3.S3
+	settings *cluster.Settings
+}
+
+var _ ExternalStorage = &s3Storage{}
+
+func makeS3Storage(
+	ctx context.Context, conf *roachpb.ExternalStorage_S3, settings *cluster.Settings,
+) (ExternalStorage, error) {
+	if conf == nil {
+		return nil, errors.Errorf("s3 upload requested but info missing")
+	}
+	region := conf.Region
+	config := conf.Keys()
+	if conf.Endpoint != "" {
+		config.Endpoint = &conf.Endpoint
+		if conf.Region == "" {
+			region = "default-region"
+		}
+		client, err := makeHTTPClient(settings)
+		if err != nil {
+			return nil, err
+		}
+		config.HTTPClient = client
+	}
+
+	// "specified": use credentials provided in URI params; error if not present.
+	// "implicit": enable SharedConfig, which loads in credentials from environment.
+	//             Detailed in https://docs.aws.amazon.com/sdk-for-go/api/aws/session/
+	// "": default to `specified`.
+	opts := session.Options{}
+	switch conf.Auth {
+	case "", authParamSpecified:
+		if conf.AccessKey == "" {
+			return nil, errors.Errorf(
+				"%s is set to '%s', but %s is not set",
+				AuthParam,
+				authParamSpecified,
+				S3AccessKeyParam,
+			)
+		}
+		if conf.Secret == "" {
+			return nil, errors.Errorf(
+				"%s is set to '%s', but %s is not set",
+				AuthParam,
+				authParamSpecified,
+				S3SecretParam,
+			)
+		}
+		opts.Config.MergeIn(config)
+	case authParamImplicit:
+		opts.SharedConfigState = session.SharedConfigEnable
+	default:
+		return nil, errors.Errorf("unsupported value %s for %s", conf.Auth, AuthParam)
+	}
+
+	sess, err := session.NewSessionWithOptions(opts)
+	if err != nil {
+		return nil, errors.Wrap(err, "new aws session")
+	}
+	if region == "" {
+		err = delayedRetry(ctx, func() error {
+			var err error
+			region, err = s3manager.GetBucketRegion(ctx, sess, conf.Bucket, "us-east-1")
+			return err
+		})
+		if err != nil {
+			return nil, errors.Wrap(err, "could not find s3 bucket's region")
+		}
+	}
+	sess.Config.Region = aws.String(region)
+	if conf.Endpoint != "" {
+		sess.Config.S3ForcePathStyle = aws.Bool(true)
+	}
+	return &s3Storage{
+		bucket:   aws.String(conf.Bucket),
+		conf:     conf,
+		prefix:   conf.Prefix,
+		s3:       s3.New(sess),
+		settings: settings,
+	}, nil
+}
+
+func (s *s3Storage) Conf() roachpb.ExternalStorage {
+	return roachpb.ExternalStorage{
+		Provider: roachpb.ExternalStorageProvider_S3,
+		S3Config: s.conf,
+	}
+}
+
+func (s *s3Storage) WriteFile(ctx context.Context, basename string, content io.ReadSeeker) error {
+	err := contextutil.RunWithTimeout(ctx, "put s3 object",
+		timeoutSetting.Get(&s.settings.SV),
+		func(ctx context.Context) error {
+			_, err := s.s3.PutObjectWithContext(ctx, &s3.PutObjectInput{
+				Bucket: s.bucket,
+				Key:    aws.String(filepath.Join(s.prefix, basename)),
+				Body:   content,
+			})
+			return err
+		})
+	return errors.Wrap(err, "failed to put s3 object")
+}
+
+func (s *s3Storage) ReadFile(ctx context.Context, basename string) (io.ReadCloser, error) {
+	// https://github.com/cockroachdb/cockroach/issues/23859
+	out, err := s.s3.GetObjectWithContext(ctx, &s3.GetObjectInput{
+		Bucket: s.bucket,
+		Key:    aws.String(filepath.Join(s.prefix, basename)),
+	})
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to get s3 object")
+	}
+	return out.Body, nil
+}
+
+func getBucketBeforeWildcard(path string) string {
+	globIndex := strings.IndexAny(path, "*?[")
+	if globIndex < 0 {
+		return path
+	}
+	return filepath.Dir(path[:globIndex])
+}
+
+func (s *s3Storage) ListFiles(ctx context.Context) ([]string, error) {
+	var fileList []string
+	baseBucket := getBucketBeforeWildcard(*s.bucket)
+
+	err := s.s3.ListObjectsPagesWithContext(
+		ctx,
+		&s3.ListObjectsInput{
+			Bucket: &baseBucket,
+		},
+		func(page *s3.ListObjectsOutput, lastPage bool) bool {
+			for _, fileObject := range page.Contents {
+				matches, err := filepath.Match(s.prefix, *fileObject.Key)
+				if err != nil {
+					continue
+				}
+				if matches {
+					s3URL := url.URL{
+						Scheme: "s3",
+						Host:   *s.bucket,
+						Path:   *fileObject.Key,
+					}
+					fileList = append(fileList, s3URL.String())
+				}
+			}
+			return !lastPage
+		},
+	)
+	if err != nil {
+		return nil, errors.Wrap(err, `failed to list s3 bucket`)
+	}
+
+	return fileList, nil
+}
+
+func (s *s3Storage) Delete(ctx context.Context, basename string) error {
+	return contextutil.RunWithTimeout(ctx, "delete s3 object",
+		timeoutSetting.Get(&s.settings.SV),
+		func(ctx context.Context) error {
+			_, err := s.s3.DeleteObjectWithContext(ctx, &s3.DeleteObjectInput{
+				Bucket: s.bucket,
+				Key:    aws.String(filepath.Join(s.prefix, basename)),
+			})
+			return err
+		})
+}
+
+func (s *s3Storage) Size(ctx context.Context, basename string) (int64, error) {
+	var out *s3.HeadObjectOutput
+	err := contextutil.RunWithTimeout(ctx, "get s3 object header",
+		timeoutSetting.Get(&s.settings.SV),
+		func(ctx context.Context) error {
+			var err error
+			out, err = s.s3.HeadObjectWithContext(ctx, &s3.HeadObjectInput{
+				Bucket: s.bucket,
+				Key:    aws.String(filepath.Join(s.prefix, basename)),
+			})
+			return err
+		})
+	if err != nil {
+		return 0, errors.Wrap(err, "failed to get s3 object headers")
+	}
+	return *out.ContentLength, nil
+}
+
+func (s *s3Storage) Close() error {
+	return nil
+}
